+++ src/SMAPI/Framework/Models/SConfig.cs
@@ -1,3 +1,6 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
 using StardewModdingAPI.Internal.ConsoleWriting;
 
 namespace StardewModdingAPI.Framework.Models
@@ -5,6 +8,35 @@ namespace StardewModdingAPI.Framework.Models
     /// <summary>The SMAPI configuration settings.</summary>
     internal class SConfig
     {
+        /********
+        ** Fields
+        ********/
+        /// <summary>The default config values, for fields that should be logged if different.</summary>
+        private static readonly IDictionary<string, object> DefaultValues = new Dictionary<string, object>
+        {
+            [nameof(CheckForUpdates)] = true,
+            [nameof(ParanoidWarnings)] =
+#if DEBUG
+                true,
+#else
+                false,
+#endif
+            [nameof(UseBetaChannel)] = Constants.ApiVersion.IsPrerelease(),
+            [nameof(GitHubProjectName)] = "Pathoschild/SMAPI",
+            [nameof(WebApiBaseUrl)] = "https://api.smapi.io",
+            [nameof(VerboseLogging)] = false,
+            [nameof(LogNetworkTraffic)] = false,
+            [nameof(DumpMetadata)] = false
+        };
+
+        /// <summary>The default values for <see cref="SuppressUpdateChecks"/>, to log changes if different.</summary>
+        private static readonly HashSet<string> DefaultSuppressUpdateChecks = new HashSet<string>(StringComparer.InvariantCultureIgnoreCase)
+        {
+            "SMAPI.ConsoleCommands",
+            "SMAPI.SaveBackup"
+        };
+
+
         /********
         ** Accessors
         ********/
@@ -15,15 +47,10 @@ namespace StardewModdingAPI.Framework.Models
         public bool CheckForUpdates { get; set; }
 
         /// <summary>Whether to add a section to the 'mod issues' list for mods which which directly use potentially sensitive .NET APIs like file or shell access.</summary>
-        public bool ParanoidWarnings { get; set; } =
-#if DEBUG
-            true;
-#else
-            false;
-#endif
+        public bool ParanoidWarnings { get; set; } = (bool)SConfig.DefaultValues[nameof(SConfig.ParanoidWarnings)];
 
         /// <summary>Whether to show beta versions as valid updates.</summary>
-        public bool UseBetaChannel { get; set; } = Constants.ApiVersion.IsPrerelease();
+        public bool UseBetaChannel { get; set; } = (bool)SConfig.DefaultValues[nameof(SConfig.UseBetaChannel)];
 
         /// <summary>SMAPI's GitHub project name, used to perform update checks.</summary>
         public string GitHubProjectName { get; set; }
@@ -45,5 +72,28 @@ namespace StardewModdingAPI.Framework.Models
 
         /// <summary>The mod IDs SMAPI should ignore when performing update checks or validating update keys.</summary>
         public string[] SuppressUpdateChecks { get; set; }
+
+
+        /********
+        ** Public methods
+        ********/
+        /// <summary>Get the settings which have been customised by the player.</summary>
+        public IDictionary<string, object> GetCustomSettings()
+        {
+            IDictionary<string, object> custom = new Dictionary<string, object>();
+
+            foreach (var pair in SConfig.DefaultValues)
+            {
+                object value = typeof(SConfig).GetProperty(pair.Key)?.GetValue(this);
+                if (!pair.Value.Equals(value))
+                    custom[pair.Key] = value;
+            }
+
+            HashSet<string> curSuppressUpdateChecks = new HashSet<string>(this.SuppressUpdateChecks ?? new string[0], StringComparer.InvariantCultureIgnoreCase);
+            if (SConfig.DefaultSuppressUpdateChecks.Count != curSuppressUpdateChecks.Count || SConfig.DefaultSuppressUpdateChecks.Any(p => !curSuppressUpdateChecks.Contains(p)))
+                custom[nameof(this.SuppressUpdateChecks)] = "[" + string.Join(", ", this.SuppressUpdateChecks ?? new string[0]) + "]";
+
+            return custom;
+        }
     }
 }